requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0

langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0

qdrant-client==1.12.0

sqlalchemy==2.0.36
psycopg[binary]==3.2.3

pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.18

# ファイルパーサー
pypdf==5.1.0
python-docx==1.1.2

# テスト
pytest==8.3.4
pytest-asyncio==0.24.0
